requirements.txt
requests
pyinstaller
pillow
pytz
